refactor to common preloader, but still keep the feature flag
