+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "bin/*",
     "crates/*",
     "example/preloader",
+    "example/common-preloader",
     "canoe/bindings",
     "canoe/provider",
     "canoe/steel/apps",
